refactor(substrate): split wasm component host by lifecycle concern
