+++ crates/aether-substrate/src/actor/wasm/component/sections.rs
@@ -1,0 +1,3 @@
+// Custom-section parsing for wasm component manifests lives in the sibling
+// `kind_manifest` module. This file is the component facade split point for
+// future section-local host helpers.
